# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CRM API Docs",
    "version": "1.0.0"
  },
  "security": [
    {
      "api_key": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crmAuthToken"
      }
    },
    "schemas": {
      "Company": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "postCode": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "vat": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "companyType": {
            "type": "string"
          },
          "contactMethods": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "email",
              "phone",
              "sms"
            ]
          },
          "brokerId": {
            "type": "string"
          }
        }
      },
...
      "UserCreated": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "roleId": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "locked"
            ],
            "createdAt": {
              "type": "string"
            },
            "updatedAt": {
              "type": "string"
            },
            "__v": {
              "type": "number"
            }
          }
        }
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/api/company/all": {
      "get": {
        "summary": "Get all companies",
        "tags": [
          "Company"
        ],
        "responses": {
          "200": {
            "description": "A list of companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/query": {
      "get": {
        "summary": "Get companies filtered by query",
        "description": "**Notes:**\n- Calling enpoint without query parameters will simply return all users, matching <b>/api/company/all</b> endpoint.\n- Query can be passed to endpoint to filter down by multiple fields.\n- Allowed fields are: _id, name, address, city, postCode, email, phone, vat, status, companyType, contactMethods, brokerId, createdAt, updatedAt\n- For name, address, email, phone, vat fields query will search for passed value within field\n- Example query: <b>?name=ACME&email=gmail.com</b>\n- No support for passing query within Swagger API Docs, use Insomnia or Postman\n",
        "tags": [
          "Company"
        ],
        "responses": {
          "200": {
            "description": "A list of companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/search/{query}": {
      "get": {
        "summary": "Get companies filtered by keyword search",
        "description": "**Notes:**\n- Calling enpoint without query parameters will simply return all users, matching <b>/api/company/all</b> endpoint.\n- Keyword search will look for passed value within multiple fields of company data, allowing for more flexible search options.\n- Allowed fields are: _id, name, address, city, postCode, email, phone, vat, status, companyType, contactMethods, brokerId, createdAt, updatedAt\n- For name, address, email, phone, vat fields query will search for passed value within field\n- Example query: <b>/api/company/search/ACME</b>\n- No support for passing query within Swagger API Docs, use Insomnia or Postman\n",
        "tags": [
          "Company"
        ],
        "responses": {
          "200": {
            "description": "A list of companies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/add": {
      "post": {
        "summary": "Create new company",
        "tags": [
          "Company"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              },
              "example": {
                "name": "ACME Inc.",
                "address": "123 The Great Plains",
                "city": "Turbulence Heights",
                "postCode": "MN 3456",
                "email": "hello@acmeinc.com",
                "phone": "+48876654432",
                "vat": "PL9878767655",
                "status": "active",
                "companyType": "lead",
                "contactMethods": [
                  "email",
                  "phone"
                ],
                "brokerId": "userID"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Company"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Company not found"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/patch/{id}": {
      "patch": {
        "summary": "Update company data",
        "tags": [
          "Company"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Company"
              },
              "example": {
                "name": "ACME Inc",
                "address": "13 Main Street",
                "city": "Warsaw",
                "postCode": "00-950",
                "email": "ashoka@gmail.com",
                "phone": "+48654543432",
                "vat": "PL8767656543",
                "status": "active",
                "companyType": "lead",
                "contactMethods": {
                  "tags": [
                    "email",
                    "phone"
                  ]
                },
                "brokerId": "userID"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/UserCreated"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/company/delete/{id}": {
      "delete": {
        "summary": "Delete company by ID",
        "tags": [
          "Company"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Company deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Company ID: 69c4f9090d2bbde75227de75 deleted"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Bad request"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "message": "Company not found"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login a user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
  },
...
}